pyproject.toml
[tool.poetry]
name = "apps"
version = "0.1.0"
description = ""
authors = ["Sergey Alekseev <alekseyserzh88@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.2"
python-dotenv = "0.12.0"
psycopg2-binary = "2.8.4"
django = "2.2.7"
djangorestframework = "3.11.0"
uwsgi = "2.0.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"